+++ src/GeneratesIdentifier.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace OG\Core\Domain;
+
+/**
+ * Indicates that an identifier can generate a value by its self.
+ */
+interface GeneratesIdentifier
+{
+    /**
+     * Generate a new Identifier.
+     *
+     * @return Identifier
+     */
+    public static function generate();
+}