+++ src/Identifier.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace OG\Core\Domain;
+
+/**
+ * An unique value to identify an entity.
+ */
+interface Identifier extends ValueObject
+{
+}